modified user collection to allow correctly the POST

+++ backendNet/Controllers/UserController.cs
@@ -19,14 +19,12 @@ namespace backendNet.Controller
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
-      Console.WriteLine("Here1");
       return Ok(await _userService.GetAllAsync().ConfigureAwait(false));
     }
 
     [HttpGet("{id}")]
     public async Task<IActionResult> Get(string id)
     {
-      Console.WriteLine("Here12");
       var user = await _userService.GetByIdAsync(id).ConfigureAwait(false);
       if (user == null)
       {
@@ -38,12 +36,15 @@ namespace backendNet.Controller
     [HttpPost]
     public async Task<IActionResult> CreateUser(User user)
     {
-        if (!ModelState.IsValid)
-        {
-            return BadRequest();
-        }
-        await _userService.CreateAsync(user).ConfigureAwait(false);
-        return Ok(user.Id);
+      if (!ModelState.IsValid)
+      {
+        return BadRequest();
+      }
+
+      user.SetPassword(user.PasswordHash);
+
+      await _userService.CreateAsync(user).ConfigureAwait(false);
+      return Ok(user.Id);
     }
 
     [HttpPut("{id:length(24)}")]
@@ -52,7 +53,7 @@ namespace backendNet.Controller
       var user = await _userService.GetByIdAsync(id).ConfigureAwait(false);
       if (user == null)
       {
-          return NotFound();
+        return NotFound();
       }
       await _userService.UpdateAsync(id, userIn).ConfigureAwait(false);
       return NoContent();
@@ -61,13 +62,21 @@ namespace backendNet.Controller
     [HttpDelete("{id:length(24)}")]
     public async Task<IActionResult> Delete(string id)
     {
-        var user = await _userService.GetByIdAsync(id).ConfigureAwait(false);
-        if (user == null)
-        {
-            return NotFound();
-        }
+     var user = await _userService.GetByIdAsync(id).ConfigureAwait(false);
+      if (user == null)
+      {
+        return NotFound();
+      }
+
+      if (user.Id != null)
+      {
         await _userService.DeleteAsync(user.Id).ConfigureAwait(false);
         return NoContent();
+      }
+      else
+      {
+        return BadRequest("User ID is null.");
+      }
     }
   }
 }
